fix: fix errors after update to new protocol

+++ src/tests/unit/game/tile/mock_box.hpp
@@ -8,8 +8,8 @@ namespace ep::tests
 {
   class MockBox: public ep::game::IBox {
   public:
-    MOCK_METHOD(double, GetX, (), (const, noexcept, override));
-    MOCK_METHOD(double, GetY, (), (const, noexcept, override));
+    MOCK_METHOD(float, GetX, (), (const, noexcept, override));
+    MOCK_METHOD(float, GetY, (), (const, noexcept, override));
     MOCK_METHOD(std::uint8_t , GetWidth, (), (const, noexcept, override));
     MOCK_METHOD(std::uint8_t , GetHeight, (), (const, noexcept, override));
   };
